\name{joeplots}
\alias{joeplots-package}
\alias{joe_palette}
\alias{joe_palettes}
\alias{joe_colors}
\alias{theme_joe}
\alias{scale_color_joe}
\alias{scale_colour_joe}
\alias{scale_fill_joe}
\alias{scale_color_joe_diverging}
\alias{scale_fill_joe_diverging}
\alias{joe_chart_size}
\alias{add_joe_footer}
\alias{save_joe}
\title{JoE chart theme and export helpers}
\description{Tools for producing consistent CWRU Journal of Economics charts.}
